Create `ClientTable`
It wraps all the operations needed for the client table. Internally
it would use the storage layer as defined by the user.

+++ simulation/src/cluster.rs
@@ -93,7 +93,7 @@ fn create_clients(
     lookup: ConnectionLookup,
     options: &ClusterOptions,
 ) -> Vec<Client<FaultyIO>> {
-    let mut clients = Vec::with_capacity(options.client_count as usize);
+    let mut clients = Vec::with_capacity(options.client_count);
 
     for client in 0..options.client_count {
         let options = ClientOptions {
